Don't block in eglot-imenu if performing non-essential task
eglot-imenu is used by imenu which in turn is used by which-func-mode
called from an idle timer.  We don't want it to block in that
situation.  Latest which-func mode now sets "non-essential" when
performing its duties, so we leverage that in eglot-imenu.

* eglot.el (eglot-imenu): Use non-essential.

GitHub-reference: close https://github.com/joaotavora/eglot/issues/212