Fix "first-differente" face in completion-pcm--hilit-commonality
Fixes: bug#42149

Depending on the position of point in the completion and the
completion style being used, it may or may not make sense for this
face to appear immediately after point.  This patch assumes that it
should appear in the first non-matched character after point, which
may likely be the next one to type to disambiguate between two or more
completions.

Suggested by Dario Gjorgjevski <dario.gjorgjevski@gmail.com>.

* lisp/minibuffer.el (completion-pcm--hilit-commonality): Fix
occasional misplacement of completions-first-differente.