Also override global flymake-diagnostic-functions
The global value of the flymake-diagnostic-functions is likely to be
of little use in Eglot-managed buffers, so don't run it.  Likely the
value flymake-proc-legacy-flymake is there which is not only likely of
little uses but also causes trouble in some situations.

The user can easily avert this by leveraging the variable
eglot-stay-out-of.

* eglot.el (eglot--managed-mode): Don't run global
flymake-diagnostic-functions.

GitHub-reference: fix https://github.com/joaotavora/eglot/issues/616