Fix frame-inner-height in non-GUI builds
Include tab bar in frame's inner height in non-GUI builds that don't
define tab-bar-height.  This is consistent with the inclusion of the
menu bar in the calculated height.  It is also consistent with TTY
frames of GUI builds, for which tab-bar-height is always zero
anyway (bug#47234).  Fix suggested by Eli Zaretskii <eliz@gnu.org>.

* lisp/frame.el (frame-inner-height): Don't assume tab-bar-height is
defined in builds --without-x.