Fix compilation warnings in non-toolkit builds
* src/xterm.c (x_create_toolkit_scroll_bar)
(x_create_horizontal_toolkit_scroll_bar): String constants for
XtSetArg are defined as const strings (in
/usr/include/X11/Xaw3d/ThreeD.h, for instance), but the argument
in XtSetArg is defined as either a const char* or a regular char*
in /usr/include/X11/Intrinsic.h.  Cast the argument to String
(which should be correct on all platforms, hopefully) to avoid a
compilation warning (bug#47452).