Fix two GTK3 event handling issues
* src/xterm.c (handle_one_xevent): For GTK3 PropertyNotify and
MapNotify events explicitly request the stored frame sizes when
the frame changes from iconified to a non-hidden state
(Bug#24526).  For Expose events do not change the frame's
visibility or iconified state.  For FocusIn events on GTK3 do
not apply the fix for Bug#42655.  The latter two changes are to
avoid that plain invisible frames get reported as iconified.