Improve first doc-string line in trivial cases
The first line of a doc-string should form a complete sentence.
Many doc-strings in this package do not follow that convention
but fixing them all would be difficult and is beyond the scope
of this commit.  However, in a few cases the fix is trivial and
this commit tackles those instances.