* lisp/printing.el (pr-global-menubar): Fix duplicate menu problem.
In Emacs 27.1 Print menu items were moved to a submenu of the File menu.
The Printing package (lisp/printing.el) replaces the Print menu (via the
function: pr-global-menubar).  The latter needs to be updated to reflect the
changes; otherwise, a second Print (sub)menu is created at the end of the
File menu.

Copyright-paperwork-exempt: yes